refactor(actions): move SaveDataset into base
one hiccup that came out of this: by moving SaveDataset down in the dep tree, the repo/test package created a circular dependency between startf and base through repo/test. By refactoring repo/test to not use base, we removed the import cycle, but now there's some fairly ugly copy-pasing going on between base.CreateDataset and repo.CreateDataset.

This problrm existed before, but seems to be a bit worse now.
