+++ base/dataset_test.go
@@ -5,13 +5,9 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/qri-io/dataset"
 	"github.com/qri-io/dataset/dstest"
-	"github.com/qri-io/ioes"
-	"github.com/qri-io/qfs"
 	"github.com/qri-io/qfs/cafs"
 	"github.com/qri-io/qri/repo"
-	"github.com/qri-io/qri/repo/profile"
 )
 
 func TestListDatasets(t *testing.T) {
@@ -68,74 +64,6 @@ func TestListDatasets(t *testing.T) {
 	}
 	if len(res) != 1 {
 		t.Error("expected one dataset with \"cit\" in their name")
-	}
-}
-
-func TestCreateDataset(t *testing.T) {
-	ctx := context.Background()
-	streams := ioes.NewDiscardIOStreams()
-	store := cafs.NewMapstore()
-	r, err := repo.NewMemRepo(testPeerProfile, store, qfs.NewMemFS(), profile.NewMemStore())
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-
-	ds := &dataset.Dataset{
-		Name:   "foo",
-		Meta:   &dataset.Meta{Title: "test"},
-		Commit: &dataset.Commit{Title: "hello"},
-		Structure: &dataset.Structure{
-			Format: "json",
-			Schema: dataset.BaseSchemaArray,
-		},
-	}
-	ds.SetBodyFile(qfs.NewMemfileBytes("body.json", []byte("[]")))
-
-	if _, err := CreateDataset(ctx, r, streams, &dataset.Dataset{}, &dataset.Dataset{}, false, true, false, true); err == nil {
-		t.Error("expected bad dataset to error")
-	}
-
-	ref, err := CreateDataset(ctx, r, streams, ds, &dataset.Dataset{}, false, true, false, true)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	refs, err := r.References(0, 10)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(refs) != 1 {
-		t.Errorf("ref length mismatch. expected 1, got: %d", len(refs))
-	}
-
-	ds.Meta.Title = "an update"
-	ds.PreviousPath = ref.Path
-	ds.SetBodyFile(qfs.NewMemfileBytes("body.json", []byte("[]")))
-
-	prev := ref.Dataset
-
-	ref, err = CreateDataset(ctx, r, streams, ds, prev, false, true, false, true)
-	if err != nil {
-		t.Fatal(err.Error())
-	}
-	refs, err = r.References(0, 10)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(refs) != 1 {
-		t.Errorf("ref length mismatch. expected 1, got: %d", len(refs))
-	}
-
-	ds.PreviousPath = ref.Path
-	ds.SetBodyFile(qfs.NewMemfileBytes("body.json", []byte("[]")))
-	prev = ref.Dataset
-
-	if ref, err = CreateDataset(ctx, r, streams, ds, prev, false, true, false, true); err == nil {
-		t.Error("expected unchanged dataset with no force flag to error")
-	}
-
-	ds.SetBodyFile(qfs.NewMemfileBytes("body.json", []byte("[]")))
-	if ref, err = CreateDataset(ctx, r, streams, ds, prev, false, true, true, true); err != nil {
-		t.Errorf("unexpected force-save error: %s", err)
 	}
 }
